Final Lab 02 Update
Replaced "dd" with "data" in the utilities.R file so the functions will take in the specified dataset versus it being hardcoded.

Reformatted the .Rmd to look like the one our Professor had in her Lab 02 tutorial

The third function looked quite complicated compared to what I briefly saw in our Professor's video. I trimmed this function and added a couple tests in the .Rmd.